LayoutTests/imported/w3c/web-platform-tests/XMLHttpRequest/abort-during-open.worker-expected.txt


FAIL Untitled assert_equals: after abort() expected 1 but got 0
